SC-027 added import for databindind, swichted size to double [EB,LC]

+++ app/src/main/java/at/tugraz05/slimcat/AddcatActivity.kt
@@ -219,7 +219,7 @@ class AddcatActivity : AppCompatActivity() {
                 0.0
             }
         }
-        binding.cat!!.size = finalSize.toInt()
+        binding.cat!!.size = finalSize.toDouble()
     }
 
     fun getWeightHintStr():String{
